refactor(website): organize and harden font previews
